+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Change History
 Version 2.0.4 (Pending)
+- Framework and library updates
+- Read version from build information instead of hardcoding it
 
 Version 2.0.3 (11/08/2022)
 - Use regular font for attendee badge names - Myriad Pro Bold is missing symbols